[models] Add django swapper #3
Closes #3
