+++ README.rst
@@ -223,6 +223,98 @@ Run tests with:
 
     ./runtests.py
 
+
+The django app `tests/openwisp2/sample_firmware_upgrader.py` adds some changes on
+top of the `openwisp-firmware-upgrader` module with the sole purpose of testing the
+module's extensibility.
+
+Extending openwisp-firmware-upgrader
+---------------------
+
+The `tests/openwisp2/sample-firmware-upgrader` may serve as an example for
+extending *openwisp-firmware-upgrader* in your own application.
+
+*openwisp-firmware-upgrader* provides a set of models and admin classes which can
+be imported, extended and reused by third party apps.
+
+To extend *openwisp-firmware-upgrader*, **you MUST NOT** add it to ``settings.INSTALLED_APPS``,
+but you must create your own app (which goes into ``settings.INSTALLED_APPS``), import the
+base classes from openwisp-firmware-upgrader and add your customizations.
+
+In order to help django find the static files and templates of *openwisp-firmware-upgrader*,
+you need to perform the steps described below.
+
+1. Install ``openwisp-utils``
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Install (and add to the requirement of your project) `openwisp-utils
+<https://github.com/openwisp/openwisp-utils>`_::
+
+    pip install openwisp-utils
+
+2. Add ``EXTENDED_APPS``
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+Add the following to your ``settings.py``:
+
+.. code-block:: python
+
+    EXTENDED_APPS = ('openwisp-firmware-upgrader',)
+
+3. Add ``openwisp_utils.staticfiles.DependencyFinder``
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Add ``openwisp_utils.staticfiles.DependencyFinder`` to
+``STATICFILES_FINDERS`` in your ``settings.py``:
+
+.. code-block:: python
+
+    STATICFILES_FINDERS = [
+        'django.contrib.staticfiles.finders.FileSystemFinder',
+        'django.contrib.staticfiles.finders.AppDirectoriesFinder',
+        'openwisp_utils.staticfiles.DependencyFinder',
+    ]
+
+4. Add ``openwisp_utils.loaders.DependencyLoader``
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Add ``openwisp_utils.loaders.DependencyLoader`` to ``TEMPLATES`` in your ``settings.py``:
+
+.. code-block:: python
+
+    TEMPLATES = [
+        {
+            'BACKEND': 'django.template.backends.django.DjangoTemplates',
+            'OPTIONS': {
+                'loaders': [
+                    'django.template.loaders.filesystem.Loader',
+                    'django.template.loaders.app_directories.Loader',
+                    'openwisp_utils.loaders.DependencyLoader',
+                ],
+                'context_processors': [
+                    'django.template.context_processors.debug',
+                    'django.template.context_processors.request',
+                    'django.contrib.auth.context_processors.auth',
+                    'django.contrib.messages.context_processors.messages',
+                ],
+            },
+        }
+    ]
+
+Extending models
+~~~~~~~~~~~~~~~~
+
+For the purpose of showing an example, we add a simple "details" field to the
+models from openwisp-firmware-upgrader [here](tests/openwisp2/sample_firmware_upgrader/models.py).
+You can add fields in a similar way in your models.py file.
+
+Extending the admin
+~~~~~~~~~~~~~~~~~~~
+
+Please checkout the sample admin.py [here](tests/openwisp2/sample_firmware_upgrader/admin.py).
+You can add changes in the `CategoryAdmin`, `BuildAdmin` and `BatchUpgradeOperationAdmin` for
+them to be reflected in your dashboard interface.
+
 Contributing
 ------------
