+++ runtests.py
@@ -11,5 +11,9 @@ if __name__ == "__main__":
     from django.core.management import execute_from_command_line
     args = sys.argv
     args.insert(1, "test")
-    args.insert(2, "openwisp_firmware_upgrader")
+    if not os.environ.get('SAMPLE_APP', False):
+        args.insert(2, "openwisp_firmware_upgrader")
+    else:
+        args.insert(2, "openwisp2")
+    # args.insert(3, "-v 3")
     execute_from_command_line(args)
