+++ tests/openwisp2/sample_firmware_upgrader/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'openwisp2.sample_firmware_upgrader.apps.SampleFirmwareUpgraderConfig'
