+++ tests/openwisp2/sample_firmware_upgrader/apps.py
@@ -1,0 +1,6 @@
+from openwisp_firmware_upgrader.apps import FirmwareUpdaterConfig
+
+
+class SampleFirmwareUpgraderConfig(FirmwareUpdaterConfig):
+    name = 'openwisp2.sample_firmware_upgrader'
+    label = 'sample_firmware_upgrader'
